% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04peqdietze-package.R
\docType{package}
\name{hw04peqdietze-package}
\alias{hw04peqdietze}
\alias{hw04peqdietze-package}
\title{hw04peqdietze: HW04 Package by eqdietze}
\description{
This package creates two codes as assigned in HW04.
}
\author{
\strong{Maintainer}: Elizabeth Dietze \email{ed0998a@student.american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
