% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot_n.R
\name{myplot_n}
\alias{myplot_n}
\title{Plot of n and output from myseq_n function}
\usage{
myplot_n(df)
}
\arguments{
\item{df}{}
}
\value{
a line plot
}
\description{
This function creates a geom_line plot where x = n and y = the output from the myseq_n function.
}
\examples{
myplot_n(my_data)
}
