% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myseq_n.R
\name{myseq_n}
\alias{myseq_n}
\title{Recursive Sequence of x and n}
\usage{
myseq_n(x, n)
}
\arguments{
\item{x}{integer}

\item{n}{integer}
}
\value{
an integer n
}
\description{
This function takes as input a vector x containing the first three numeric elements of this sequence and a positive integer n denoting the final element of the sequence to calculate and returns element n.
}
\examples{
myseq_n(x = c(2, 3, 3), n = 3)
}
